requirements.txt
# Email Processing System Requirements
# For GPU-accelerated processing with RTX 5080

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# GPU Acceleration (CuPy for CUDA 12.x)
cupy-cuda12x>=13.0.0

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
chardet>=5.0.0
psutil>=5.9.0
